moved location for entities
