--- src/common/entities/unlock-token.ts
@@ -1,11 +1,0 @@
-import { UnlockMileStoneModel } from "./unlock-schedule";
-
-export interface ILockedTokens {
-  xmex: string | boolean,
-  lkmex: string | boolean;
-}
-
-export interface IUnlockFields {
-  unlockEpoch?: number,
-  unlockSchedule?: UnlockMileStoneModel[];
-}
