+++ src/common/locked-asset/model/locked.tokens.interface.ts
@@ -1,0 +1,4 @@
+export interface LockedTokensInterface {
+  xmex: string,
+  lkmex: string;
+}
