+++ src/common/locked-asset/model/unlock.fields.interface.ts
@@ -1,0 +1,6 @@
+import { UnlockMileStoneModel } from "./unlock.milestone.model";
+
+export interface UnlockFieldsInterface {
+  unlockEpoch?: number,
+  unlockSchedule?: UnlockMileStoneModel[];
+}
